Adds documentation
